Add support for Bundler 1.x and Bundle 2.0 (#77) [semver:minor]
* Add support for Bundler 1.x and Bundle 2.0

To fix a deprecation warning about the use of `--path` in May a change
was shipped to use `bundle config set path`. The `bundle config` command
was changed in 2.1 to use `bundle config [set|list|get|unset] <name> <val>`
but before then the format was `bundle config <name> <val>`. Following on
from that then, anyone using this orb with Bundler 1.x or 2.0 will have a
bundle config setting called `set` and the value of it will be
`path <<parameter.path>>`. This simply doesn't work and results in gems
being installed into ~/rubygems outside the project directory and
consequently a completely empty cache during the save step.

Running `bundle config set` without anything else throws an error on
Bundler 2.1 because it expects `name` and `value` to be supplied. On
the older Bundler versions it succeeds and returns the setting for
config variable called `set`. We can use this fact to implement a
Bundler semantics check and issue the `path` and `deployment` settings
correctly for the currently used Bundler binary.

* Update Linux machine image & sourcing

Co-authored-by: Ricardo N Feliciano <FelicianoTech@gmail.com>