Give warning if rbenv is installed when installing
Installing ruby with this orb will install via RVM, which can sometimes
conflict with rbenv if both are installed.

In particular, we had a customer who was experiencing issues with our
latest machine images that support rbenv.  The customer had a
.ruby-version file in their home directory which was breaking
installation of ruby via rvm when the system rbenv was already present.

To this end if we find you have rbenv installed and also have a
.ruby-version file present in your working directory we will emit a
warning suggesting that instead of installing via the orb you simply
can use "rbenv install" directly.

We made the decision to only emit a warning rather than an error and
halting the build because this orb is also used on numerous customer
docker images where we cannot entirely know the exact configuration of
that image and we want to avoid inadvertently unnecessary breaking a
build that may in fact be functioning correctly.